# Conflicts:
#	backend/src/main/java/com/backend/comment/controller/CommentController.java
#	backend/src/main/java/com/backend/comment/entity/Comment.java
#	backend/src/main/java/com/backend/comment/repository/CommentRepository.java
#	backend/src/main/java/com/backend/comment/service/CommentService.java
#	backend/src/main/java/com/backend/comment/service/CommentServiceImpl.java
#	backend/src/main/java/com/backend/post/dto/PostCreateRequestDto.java
#	backend/src/main/java/com/backend/post/dto/PostUpdateRequestDto.java
#	backend/src/main/java/com/backend/post/entity/Post.java
#	backend/src/main/java/com/backend/post/service/PostService.java
#	backend/src/main/java/com/backend/role/entity/Role.java
#	backend/src/main/java/com/backend/role/entity/RoleEnum.java
#	backend/src/main/java/com/backend/user/entity/User.java

+++ backend/src/main/java/com/backend/role/entity/Role.java
@@ -12,7 +12,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+    private Long id;
 
     @Column(nullable = false)
     private RoleEnum role;
